+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+## [4.6.1] - 2023-10-25
+
+### Changed
+
+- All e2e tests are now run using Playwright rather than RTL.
+- Application is now built using Vite rather than Webpack.
+
 ## [4.6.0] - 2023-10-23
 
 ### Added
@@ -475,7 +482,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added auto-scroll to TX pane, closes #89.
 - Added special delete behaviour for backspace button when in "send on enter" mode, closes #90.
 
-[unreleased]: https://github.com/gbmhunter/NinjaTerm/compare/v4.6.0...HEAD
+[unreleased]: https://github.com/gbmhunter/NinjaTerm/compare/v4.6.1...HEAD
+[4.6.1]: https://github.com/gbmhunter/NinjaTerm/compare/v4.6.0...v4.6.1
 [4.6.0]: https://github.com/gbmhunter/NinjaTerm/compare/v4.5.2...v4.6.0
 [4.5.2]: https://github.com/gbmhunter/NinjaTerm/compare/v4.5.1...v4.5.2
 [4.5.1]: https://github.com/gbmhunter/NinjaTerm/compare/v4.5.0...v4.5.1